Support multiple Memory objects per MemorySection, for consistency
Other sections (exports, tables, etc) allow for multiple objects per
section.  The current WASM 1.1 spec does not support multiple memories,
but the inclusion of a vector of memories implies this might be possible in
the future.  Regardless, recognize/parse multiple Memory objects per
MemorySection, for consistency with the other sections.